Update CodeLink 04.py
Corrected the playRandomly() function to test the board before the queen is placed.  The old version tested after placement, which caused every board to be rejected due to the very queen being placed.  This did not matter for the demo code, but might have made it harder to debug the CodeLink challenge.